pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tth"
version = "0.1.0"
description = "Text-to-Human video system with emotion and character controllability"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "websockets>=12",
  "structlog>=24",
  "pyyaml>=6",
  "python-dotenv>=1",
  "pillow>=12.1.1",
  "numpy>=2.4.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "pytest-httpx>=0.30",
  "mypy>=1.10",
  "ruff>=0.4",
]
train = [
  "torch",
  "pytorch-lightning",
  "omegaconf",
  "hydra-core",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tth"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
